requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==1.6.1
pymongo==4.5.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
Werkzeug==2.3.7
